backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pymysql==1.1.1
cryptography==42.0.7          # pymysql TLS support

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Rate limiting
slowapi==0.1.9

# HTTP client (for Stripe webhook verification etc.)
httpx==0.27.0

# Payment
stripe==9.5.0

# Logging
structlog==24.1.0

# CORS, middleware
python-multipart==0.0.9

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0                  # async test client
factory-boy==3.3.0
faker==25.2.0

# Dev
python-dotenv==1.0.1
